+++ src/export/exported-tests.c
@@ -5,6 +5,12 @@ sexp* rlang_r_string(sexp* str) {
 }
 
 
+// chr.c
+
+sexp* rlang_test_nms_are_duplicated(sexp* nms, sexp* from_last) {
+  return r_nms_are_duplicated(nms, r_lgl_get(from_last, 0));
+}
+
 // cnd.c
 
 sexp* rlang_test_r_warn(sexp* x) {