+++ src/lib/vec-chr.c
@@ -99,10 +99,33 @@ sexp* chr_append(sexp* chr, sexp* r_str) {
   return out;
 }
 
+sexp* r_nms_are_duplicated(sexp* nms, bool from_last) {
+  if (r_typeof(nms) != r_type_character) {
+    r_abort("Internal error: Expected a character vector of names for checking duplication");
+  }
+  sexp* dups = KEEP(Rf_duplicated(nms, from_last));
+
+  r_ssize n = r_length(dups);
+  int* dups_ptr = r_lgl_deref(dups);
+  sexp** nms_ptr = r_chr_deref(nms);
+
+  for (r_ssize i = 0; i < n; ++i, ++dups_ptr, ++nms_ptr) {
+    if (*nms_ptr == r_empty_str || *nms_ptr == r_missing_str) {
+      *dups_ptr = false;
+    }
+  }
+
+  FREE(1);
+  return dups;
+}
+
 
 sexp* r_shared_empty_chr = NULL;
+sexp* r_empty_str = NULL;
 
 void r_init_library_vec_chr() {
   r_shared_empty_chr = r_chr("");
   r_mark_precious(r_shared_empty_chr);
+
+  r_empty_str = r_chr_get(r_shared_empty_chr, 0);
 }