+++ tests/testthat/helper-c-api.R
@@ -1,0 +1,4 @@
+
+nms_are_duplicated <- function(nms, from_last = FALSE) {
+  .Call(rlang_nms_are_duplicated, nms, from_last)
+}